pyproject.toml
[project]
name = "realtime-streaming-pipeline"
version = "0.1.0"
description = "A production-ready real-time data streaming pipeline using Apache Kafka, Python, and PostgreSQL"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Kafka
    "confluent-kafka==2.3.0",
    "kafka-python==2.0.2",
    
    # Data Processing
    "pandas==2.1.3",
    "numpy==1.26.2",
    "python-dateutil==2.8.2",
    
    # Database
    "psycopg2-binary==2.9.9",
    "sqlalchemy==2.0.23",
    "influxdb-client==1.38.0",
    
    # Configuration
    "pyyaml==6.0.1",
    "python-dotenv==1.0.0",
    
    # Monitoring & Logging
    "prometheus-client==0.19.0",
    "structlog==23.2.0",
    
    # Data Validation
    "pydantic==2.5.2",
    "jsonschema==4.20.0",
    
    # Serialization
    "avro-python3==1.10.2",
    "fastavro==1.9.0",
    
    # API & HTTP
    "requests==2.31.0",
    "aiohttp==3.9.1",
    
    # Utilities
    "click==8.1.7",
    "rich==13.7.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "faker==20.1.0",
    
    # Linting & Formatting
    "flake8>=6.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["producers", "consumers", "utils", "config"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=term-missing"
